MVC kindof
