+++ Platformer/Levels/1.level
@@ -12,7 +12,7 @@ g                                            gg
 g    gg
 g              c c                          
 g             c c c                            
-g            c c c c                             F
+g            c c c c                               F
 g           w                   f                 gg
 g   P  wg       w  e       w         w        gg       gggggg
 gggggggggg   gggggggggggggggggggggg g  gg gg
