+++ Platformer/Levels/2.level
@@ -6,5 +6,5 @@
                ggggg
 			E
            gggg
-	 F
+ sss F
 gggggggg
