+++ Platformer/Levels/3.level
@@ -5,5 +5,5 @@
 	     w                    gg
  P      w  c             gg
 ggggggggggggg   gg   gg             gg
-                             gg     
+                             gg   l  
                                  ggg
