+++ Platformer/MainWindow.xaml.cs
@@ -14,6 +14,7 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using Platformer.Views;
 using Platformer.Models;
+using Platformer.Data;
 
 namespace Platformer
 {
