+++ Platformer/Models/Enemy.cs
@@ -1,0 +1,72 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Windows;
+
+namespace Platformer.Models
+{
+    class Enemy : Actor
+    {
+        protected Enemy(double x, double y, int EnemyWidth, int EnemyHeight) : base(x, y, EnemyWidth, EnemyHeight)
+        {
+            ;
+        }
+
+        public void TurnAround()
+        {
+            Velocity = -Velocity;
+        }
+
+        public void StopMoving()
+        {
+            Velocity = 0;
+        }
+
+        public void StartMoving()
+        {
+            if (this is SmallEnemy)
+            {
+                Velocity = 3;
+            }
+            else if (this is BigEnemy)
+            {
+                Velocity = 1;
+            }
+            else if (this is FlyingEnemy)
+            {
+                Velocity = 2;
+            }
+
+        }
+    }
+
+    class SmallEnemy : Enemy
+    {
+        public SmallEnemy(double x, double y) : base(x, y, Config.SmallEnemyWidth, Config.SmallEnemyHeight)
+        {
+            hitsToKill = 1;
+            Velocity = 3;
+            Brush = Config.SmallEnemyBrush;
+        }
+    }
+
+    class BigEnemy : Enemy
+    {
+        public BigEnemy(double x, double y) : base(x, y, Config.BigEnemyWidth, Config.BigEnemyHeight)
+        {
+            hitsToKill = 2;
+            Velocity = 1;
+            Brush = Config.BigEnemyBrush;
+        }
+    }
+
+    class FlyingEnemy : Enemy
+    {
+        public FlyingEnemy(double x, double y) : base(x, y, Config.UnitWidth, Config.UnitHeight)
+        {
+            hitsToKill = 1;
+            Velocity = 2;
+            Brush = Config.FlyingEnemy1;
+        }
+    }
+}
