--- Platformer/Models/PlatformerContext.cs
@@ -1,50 +1,0 @@
-﻿using System;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata;
-
-#nullable disable
-
-namespace Platformer.Models
-{
-    public partial class PlatformerContext : DbContext
-    {
-        public PlatformerContext()
-        {
-        }
-
-        public PlatformerContext(DbContextOptions<PlatformerContext> options)
-            : base(options)
-        {
-        }
-
-        public virtual DbSet<LeaderboardEntry> LeaderboardEntries { get; set; }
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            if (!optionsBuilder.IsConfigured)
-            {
-                optionsBuilder.UseSqlServer("Data Source = (LocalDB)\\MSSQLLocalDB; AttachDbFilename = |DataDirectory|\\PlatformerDatabase.mdf; Integrated Security = True");
-            }
-        }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.HasAnnotation("Relational:Collation", "SQL_Latin1_General_CP1_CI_AS");
-
-            modelBuilder.Entity<LeaderboardEntry>(entity =>
-            {
-                entity.Property(e => e.Name)
-                    .IsRequired()
-                    .HasMaxLength(50);
-
-                entity.Property(e => e.Time)
-                    .IsRequired()
-                    .HasMaxLength(50);
-            });
-
-            OnModelCreatingPartial(modelBuilder);
-        }
-
-        partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
-    }
-}
